chess/requirements.txt
asgiref==3.5.2 ##auto
chess==1.9.3
Django==4.1.3 ##framework
djangorestframework==3.14.0 ##create api
pytz==2022.6 ## auto
sqlparse==0.4.3 ##auto